change batches to batch builders where necessary and prepare for batches page
